+++ src/compiler/codegen/index.ts
@@ -259,7 +259,7 @@ export function genFor(
     state.warn(
       `<${el.tag} v-for="${alias} in ${exp}">: component lists rendered with ` +
         `v-for should have explicit keys. ` +
-        `See https://vuejs.org/guide/list.html#key for more info.`,
+        `See https://v2.vuejs.org/v2/guide/list.html#key for more info.`,
       el.rawAttrsMap['v-for'],
       true /* tip */
     )